+++ apps/tergite-mss/api/rest/app_kwargs.py
@@ -26,8 +26,8 @@ from services.auth import service as auth_service
 from services.devices import service as devices_service
 from services.external import bcc, puhuri
 
-from .dependencies import get_default_mongodb
 from .routers.devices import ws_manager as devices_ws_manager
+from .utils import get_default_mongodb
 
 
 def get_app_kwargs() -> Dict[str, Any]: