+++ apps/tergite-mss/api/rest/utils.py
@@ -1,6 +1,7 @@
 # This code is part of Tergite
 #
-# (C) Copyright Martin Ahindura 2024
+# (C) Copyright Martin Ahindura 2023, 2024
+# (C) Copyright Chalmers Next Labs AB 2026
 #
 # This code is licensed under the Apache License, Version 2.0. You may
 # obtain a copy of this license in the LICENSE.txt file in the root directory
@@ -10,57 +11,24 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-"""Some utility functions for the routers"""
+"""Miscellaneous utility functions for the routers"""
+
 import logging
 import uuid
 from datetime import datetime
+from typing import Dict
 
 from cryptography.exceptions import InvalidSignature
-from fastapi import WebSocketException
-from fastapi.middleware.cors import CORSMiddleware
-from starlette.datastructures import Headers, MutableHeaders
+from fastapi import Depends, WebSocketException
 from starlette.requests import Request
 from starlette.status import WS_1008_POLICY_VIOLATION
-from starlette.types import Message, Send
 from starlette.websockets import WebSocket
 
 import settings
-from utils.api import (
-    GeneralMessage,
-    WebsocketRequestLog,
-    get_request_logs_store,
-    verify_ws_signature,
-)
-from utils.crypto import get_uuid4_str
-from utils.exc import InvalidRequestIDError
-from utils.logging import access_logger
-
-
-class TergiteCORSMiddleware(CORSMiddleware):
-    """Custom CORS middleware to handle Tergite specific behaviour"""
-
-    async def send(
-        self, message: Message, send: Send, request_headers: Headers
-    ) -> None:
-        if message["type"] != "http.response.start":
-            await send(message)
-            return
-
-        message.setdefault("headers", [])
-        headers = MutableHeaders(scope=message)
-        headers.update(self.simple_headers)
-        origin = request_headers["Origin"]
-
-        # If all origins are allowed, respond back with the Origin header
-        if self.allow_all_origins:
-            self.allow_explicit_origin(headers, origin)
-
-        # If we only allow specific origins, then we have to mirror back
-        # the Origin header in the response.
-        elif not self.allow_all_origins and self.is_allowed_origin(origin=origin):
-            self.allow_explicit_origin(headers, origin)
-
-        await send(message)
+from services.external import bcc
+from utils.api import get_request_logs_store, verify_ws_signature
+from utils.exc import InvalidRequestIDError, UnknownBccError
+from utils.mongodb import get_mongodb
 
 
 async def get_request_id(request: Request) -> str:
@@ -85,52 +53,6 @@ async def get_request_id(request: Request) -> str:
         return request_id
     except AttributeError:
         raise InvalidRequestIDError(f"The request ID '' is not a valid UUID4")
-
-
-class WebsocketConnectionManager:
-    """Manages connections to a websockets endpoint"""
-
-    def __init__(self):
-        self.active_connections: list[WebSocket] = []
-
-    async def connect(self, websocket: WebSocket):
-        """Allows a connection to be established"""
-        request_id = websocket.headers.get("x-request-id")
-        if request_id is None:
-            request_id = get_uuid4_str()
-
-        websocket.state.request_id = request_id
-        requests_store = get_request_logs_store()
-        if not requests_store.exists(request_id):
-            requests_store.insert(WebsocketRequestLog.from_request(websocket))
-
-        await websocket.accept(headers=[(b"x-request-id", request_id.encode("utf8"))])
-        self.active_connections.append(websocket)
-        access_logger.info(f"Connected: {websocket.client.host} at {websocket.url}")
-
-    def disconnect(self, websocket: WebSocket):
-        """Removes a given websocket connection"""
-        self.active_connections.remove(websocket)
-        access_logger.info(f"Disconnected: {websocket.client.host} at {websocket.url}")
-
-    async def reply(self, websocket: WebSocket, message: GeneralMessage):
-        """Sends a message to the given websocket"""
-        await websocket.send_json(message)
-
-    async def broadcast(self, message: str):
-        """Broadcasts a message to all connections on a websocket endpoint"""
-        for connection in self.active_connections:
-            await connection.send_text(message)
-
-    async def close_all(self, reason: str = "Server closing connection"):
-        """Closes all active connections
-
-        Args:
-            reason: Reason for closing the connection
-        """
-        for connection in list(self.active_connections):
-            access_logger.info(f"Closing: {connection.client.host} at {connection.url}")
-            await connection.close(code=1001, reason=reason)
 
 
 def get_verified_device_name(websocket: WebSocket) -> str:
@@ -203,3 +125,31 @@ def _is_valid_uuid4(value):
     except (ValueError, TypeError):
         return False
     return str(temp_uuid) == value
+
+
+async def get_default_mongodb():
+    return get_mongodb(
+        url=f"{settings.CONFIG.database.url}", name=settings.CONFIG.database.name
+    )
+
+
+async def get_bcc_client(
+    backend: str,
+    bcc_clients_map: Dict[str, bcc.BccClient] = Depends(bcc.get_client_map),
+) -> bcc.BccClient:
+    """Dependency injector to return the BCC client
+
+    Args:
+        backend (str): the name of the backend
+        bcc_clients_map (BccClientsMap): the map of BCC clients
+
+    Returns:
+        bcc.BccClient: the BCC client
+
+    Raises:
+        UnknownBccError: Unknown backend '{backend}'
+    """
+    try:
+        return bcc_clients_map[backend]
+    except KeyError:
+        raise UnknownBccError(f"Unknown backend '{backend}'")