+++ apps/tergite-mss/tests/api/rest/test_jobs.py
@@ -22,6 +22,7 @@ from pytest_mock import MockerFixture
 
 from services.auth import Project
 from tests._utils.auth import TEST_PROJECT_EXT_ID, TEST_USER_ID, get_db_record
+from tests._utils.bcc import create_bcc_headers
 from tests._utils.date_time import (
     get_current_timestamp_str,
 )
@@ -85,6 +86,14 @@ _PAGINATE_AND_SEARCH_PARAMS = [
     (skip, limit, sort, search)
     for skip, limit, sort in _SKIP_LIMIT_SORT_PARAMS
     for search in _SEARCH_PARAMS
+]
+_DEVICES_AND_JOB_UPDATES = [
+    (device, job_update) for device in _DEVICES[:2] for job_update in _JOB_UPDATES
+]
+_DEVICES_AND_JOB_TIMESTAMPED_UPDATES = [
+    (device, job_update)
+    for device in _DEVICES[:2]
+    for job_update in _JOB_TIMESTAMPED_UPDATES
 ]
 
 
@@ -257,9 +266,9 @@ def test_find_jobs(
         assert got == expected
 
 
-@pytest.mark.parametrize("raw_payload", _JOB_UPDATES)
-def test_update_job(db, client, raw_payload: dict, app_token_header, freezer):
-    """PUT to /jobs/{job_id} updates the job with the given object, it ignores job_id"""
+@pytest.mark.parametrize("device, raw_payload", _DEVICES_AND_JOB_UPDATES)
+def test_update_job(db, client, device: str, raw_payload: dict, freezer):
+    """Pushing 'job_updated'-DeviceEvent to /devices/ws/{name} updates the job with the given object"""
     job_id = raw_payload["job_id"]
     raw_jobs = with_incremental_timestamps(
         _JOBS_LIST, fields=["created_at", "calibration_date"]
@@ -274,17 +283,22 @@ def test_update_job(db, client, raw_payload: dict, app_token_header, freezer):
 
     insert_in_collection(database=db, collection_name=_COLLECTION, data=raw_jobs)
 
-    ignored_fields = ["unexpected_field", "random_field", "job_id"]
-    payload = {**raw_payload, "job_id": f"{uuid.uuid4()}"}
+    ignored_fields = ["unexpected_field", "random_field"]
+
+    url = f"/devices/ws/{device}"
+    headers = create_bcc_headers(device)
 
     # using context manager to ensure on_startup runs
-    with client as client:
-        response = client.put(
-            f"/jobs/{job_id}",
-            json={**payload},
-            headers=app_token_header,
+    with client.websocket_connect(url, headers=headers) as client:
+        client.send_json(
+            {
+                "name": "job_updated",
+                "data": raw_payload,
+            }
         )
-        got = response.json()
+        response = client.receive_json()
+        got = response["data"]
+
         actual_update, _ = prune(raw_payload, ignored_fields)
         expected_job.update(
             actual_update,
@@ -300,16 +314,16 @@ def test_update_job(db, client, raw_payload: dict, app_token_header, freezer):
             _filter={"job_id": job_id},
         )[0]
 
-        assert response.status_code == 200
+        assert response["status"] == "success"
         assert got == expected_response
         assert job_after_update == expected_job
 
 
-@pytest.mark.parametrize("raw_payload", _JOB_TIMESTAMPED_UPDATES)
+@pytest.mark.parametrize("device, raw_payload", _DEVICES_AND_JOB_TIMESTAMPED_UPDATES)
 def test_update_job_resource_usage(
-    db, client, project_id, raw_payload: dict, app_token_header, freezer
+    db, client, project_id, device: str, raw_payload: dict, freezer
 ):
-    """PUT to /jobs/{job_id} updates the job's resource usage if passed a payload with "timestamps" property"""
+    """Pushing 'job_updated' event to /devices/ws/{name} updates the job's resource usage if passed with 'timestamps' field"""
     raw_jobs = with_incremental_timestamps(
         _JOBS_LIST, fields=["created_at", "calibration_date"]
     )
@@ -323,16 +337,19 @@ def test_update_job_resource_usage(
         db, schema=Project, _filter={"ext_id": TEST_PROJECT_EXT_ID}
     )
 
+    url = f"/devices/ws/{device}"
+    headers = create_bcc_headers(device)
+
     # using context manager to ensure on_startup runs
-    with client as client:
-        # check for idempotence
-        for _ in range(3):
-            response = client.put(
-                f"/jobs/{job_id}",
-                json=raw_payload,
-                headers=app_token_header,
-            )
-            assert response.status_code == 200
+    with client.websocket_connect(url, headers=headers) as client:
+        client.send_json(
+            {
+                "name": "job_updated",
+                "data": raw_payload,
+            }
+        )
+        response = client.receive_json()
+        assert response["status"] == "success"
 
     project_after_update = get_db_record(
         db, schema=Project, _filter={"ext_id": TEST_PROJECT_EXT_ID}
@@ -345,33 +362,35 @@ def test_update_job_resource_usage(
     assert actual_resource_usage == expected_resource_usage
 
 
-@pytest.mark.parametrize("payload", _JOB_TIMESTAMPED_UPDATES)
+@pytest.mark.parametrize("device, payload", _DEVICES_AND_JOB_TIMESTAMPED_UPDATES)
 def test_update_job_resource_usage_advanced(
-    db, client, project_id, payload: dict, app_token_header, freezer
+    db, client, project_id, device: str, payload: dict, freezer
 ):
-    """PUT to /jobs/{job_id} updates the job's resource usage if passed a payload with "timestamps.execution" field"""
+    """Pushing 'job_updated' event to /devices/ws/{name} updates the job's resource usage if passed with "timestamps.execution" field"""
     raw_jobs = with_incremental_timestamps(
         _JOBS_LIST, fields=["created_at", "calibration_date"]
     )
     raw_jobs = with_current_timestamps(raw_jobs, fields=["updated_at"])
     job_list = [{**item, "project_id": str(project_id)} for item in raw_jobs]
     insert_in_collection(database=db, collection_name=_COLLECTION, data=job_list)
-    job_id = payload["job_id"]
 
     project_before_update = get_db_record(
         db, schema=Project, _filter={"ext_id": TEST_PROJECT_EXT_ID}
     )
 
+    url = f"/devices/ws/{device}"
+    headers = create_bcc_headers(device)
+
     # using context manager to ensure on_startup runs
-    with client as client:
-        # check for idempotence
-        for _ in range(3):
-            response = client.put(
-                f"/jobs/{job_id}",
-                json=payload,
-                headers=app_token_header,
-            )
-            assert response.status_code == 200
+    with client.websocket_connect(url, headers=headers) as client:
+        client.send_json(
+            {
+                "name": "job_updated",
+                "data": payload,
+            }
+        )
+        response = client.receive_json()
+        assert response["status"] == "success"
 
     project_after_update = get_db_record(
         db, schema=Project, _filter={"ext_id": TEST_PROJECT_EXT_ID}